+++ Cargo.toml
@@ -10,7 +10,10 @@ edition = "2021"
 rust-version = "1.70"
 
 [features]
-asio = ["asio-sys", "num-traits"] # Only available on Windows. See README for setup instructions.
+asio = [
+    "asio-sys",
+    "num-traits",
+] # Only available on Windows. See README for setup instructions.
 
 # Deprecated, the `oboe` backend has been removed
 oboe-shared-stdcxx = []
@@ -36,8 +39,8 @@ windows = { version = "0.54.0", features = [
     "Win32_System_SystemServices",
     "Win32_System_Variant",
     "Win32_Media_Multimedia",
-    "Win32_UI_Shell_PropertiesSystem"
-]}
+    "Win32_UI_Shell_PropertiesSystem",
+] }
 audio_thread_priority = { version = "0.33.0", optional = true }
 asio-sys = { version = "0.2", path = "asio-sys", optional = true }
 num-traits = { version = "0.2.6", optional = true }
@@ -60,6 +63,8 @@ objc2-core-audio = { version = "0.3.1", default-features = false, features = [
     "std",
     "AudioHardware",
     "AudioHardwareDeprecated",
+    "objc2",
+    "objc2-foundation",
 ] }
 objc2-audio-toolbox = { version = "0.3.1", default-features = false, features = [
     "std",
@@ -70,20 +75,44 @@ objc2-core-audio-types = { version = "0.3.1", default-features = false, features
     "std",
     "CoreAudioBaseTypes",
 ] }
+objc2-core-foundation = { version = "0.3.1" }
+objc2-foundation = { version = "0.3.1" }
+objc2 = { version = "0.6.2" }
 
 [target.'cfg(target_os = "emscripten")'.dependencies]
 wasm-bindgen = { version = "0.2.89" }
 wasm-bindgen-futures = "0.4.33"
 js-sys = { version = "0.3.35" }
-web-sys = { version = "0.3.35", features = [ "AudioContext", "AudioContextOptions", "AudioBuffer", "AudioBufferSourceNode", "AudioNode",  "AudioDestinationNode", "Window", "AudioContextState"] }
+web-sys = { version = "0.3.35", features = [
+    "AudioContext",
+    "AudioContextOptions",
+    "AudioBuffer",
+    "AudioBufferSourceNode",
+    "AudioNode",
+    "AudioDestinationNode",
+    "Window",
+    "AudioContextState",
+] }
 
 [target.'cfg(all(target_arch = "wasm32", target_os = "unknown"))'.dependencies]
 wasm-bindgen = { version = "0.2.58", optional = true }
 js-sys = { version = "0.3.35" }
-web-sys = { version = "0.3.35", features = [ "AudioContext", "AudioContextOptions", "AudioBuffer", "AudioBufferSourceNode", "AudioNode",  "AudioDestinationNode", "Window", "AudioContextState"] }
+web-sys = { version = "0.3.35", features = [
+    "AudioContext",
+    "AudioContextOptions",
+    "AudioBuffer",
+    "AudioBufferSourceNode",
+    "AudioNode",
+    "AudioDestinationNode",
+    "Window",
+    "AudioContextState",
+] }
 
 [target.'cfg(target_os = "android")'.dependencies]
-ndk = { version = "0.9", default-features = false, features = ["audio", "api-level-26"]}
+ndk = { version = "0.9", default-features = false, features = [
+    "audio",
+    "api-level-26",
+] }
 ndk-context = "0.1"
 jni = "0.21"
 num-derive = "0.4"
